+++ src/services/ChargeService.php
@@ -12,12 +12,14 @@ namespace lukeyouell\stripecheckout\services;
 
 use lukeyouell\stripecheckout\StripeCheckout;
 use lukeyouell\stripecheckout\elements\Charge;
+use lukeyouell\stripecheckout\elements\db\ChargeQuery;
 
 use Craft;
 use craft\base\Component;
 use craft\helpers\Json;
 
 use yii\base\Exception;
+use yii\web\NotFoundHttpException;
 
 class ChargeService extends Component
 {
@@ -28,7 +30,7 @@ class ChargeService extends Component
     {
         $response = $this->createStripeCharge($request);
 
-        if (!$response['success']) {
+        if (!isset($response['charge'])) {
             return $response;
         }
 
@@ -59,15 +61,15 @@ class ChargeService extends Component
                 'shipping'      => $request['shipping'],
                 'metadata'      => $request['metadata'],
             ]);
-
-            $response['success'] = true;
         } catch (\Stripe\Error\Base $e) {
             Craft::$app->getErrorHandler()->logException($e);
-            $response['success'] = false;
+
+            $body = $e->getJsonBody();
+            $response = $body['error'];
+
             $response['message'] = $e->getMessage();
         } catch (Exception $e) {
             Craft::$app->getErrorHandler()->logException($e);
-            $response['success'] = false;
             $response['message'] = $e->getMessage();
         }
 
@@ -78,7 +80,7 @@ class ChargeService extends Component
     {
         if ($stripeCharge) {
           $charge = new Charge();
-          
+
           $charge->stripeId       = $stripeCharge->id ?? null;
           $charge->email          = $stripeCharge->receipt_email ?? null;
           $charge->live           = $stripeCharge->livemode ?? null;
@@ -102,6 +104,62 @@ class ChargeService extends Component
           if ($res) {
               return $charge;
           }
+        }
+
+        return null;
+    }
+
+    public function updateCharge($stripeCharge = null)
+    {
+        if ($stripeCharge) {
+          $charge = $this->getChargeByStripeId($stripeCharge->id);
+
+          $charge->email          = $stripeCharge->receipt_email ?? null;
+          $charge->live           = $stripeCharge->livemode ?? null;
+          $charge->chargeStatus   = $stripeCharge->status ?? null;
+          $charge->paid           = $stripeCharge->paid ?? null;
+          $charge->refunded       = $stripeCharge->refunded ?? null;
+          $charge->amount         = $stripeCharge->amount ?? null;
+          $charge->amountRefunded = $stripeCharge->amount_refunded ?? null;
+          $charge->currency       = $stripeCharge->currency ?? null;
+          $charge->description    = $stripeCharge->description ?? null;
+          $charge->source         = $stripeCharge->source ? Json::encode($stripeCharge->source) : null;
+          $charge->refunds        = $stripeCharge->refunds ? Json::encode($stripeCharge->refunds->data) : null;
+          $charge->shipping       = $stripeCharge->shipping ? Json::encode($stripeCharge->shipping) : null;
+          $charge->metadata       = $stripeCharge->metadata ? Json::encode($stripeCharge->metadata) : null;
+          $charge->outcome        = $stripeCharge->outcome ? Json::encode($stripeCharge->outcome) : null;
+          $charge->failureCode    = $stripeCharge->failure_code ?? null;
+          $charge->failureMessage = $stripeCharge->failure_message ?? null;
+
+          $res = Craft::$app->getElements()->saveElement($charge, true, false);
+
+          if ($res) {
+              return $charge;
+          }
+        }
+
+        return null;
+    }
+
+    public function getChargeById($id = null)
+    {
+        if ($id) {
+            $query = new ChargeQuery(Charge::class);
+            $query->id = $id;
+
+            return $query->one();
+        }
+
+        return null;
+    }
+
+    public function getChargeByStripeId($id = null)
+    {
+        if ($id) {
+            $query = new ChargeQuery(Charge::class);
+            $query->stripeId = $id;
+
+            return $query->one();
         }
 
         return null;