+++ CHANGELOG.md
@@ -6,4 +6,13 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Default templates
+- The ability to filter, search, sort and delete charges
+- The ability to override the plugin name
+
 ### Changed
+
+- Updated Stripe to `^6.7`
+- Charges are now elements
+- Improved error handling
+- Improved settings templates